Added buttons and time logging
Added 'Time Tracking' buttons, 'start' and 'stop', that keep track of when the users starts working on that task and when they stop working on the task. These times track when the person started their task in a format of 'mm-dd-yyyy'. Added a 'Time Logs' section. The section states at what time the user pressed the start button and when the users pressed the stop button. The times of when they started and stopped are used to tell the user of how long they worked for, during each session. This is expressed in message stating 'You worked for: xx:yy:zz'.