Progression bar & Estimated Hours
Finalized the progression bar that displays on the homescreen. Users, once they create the task and set their hours, will be able to see a visual representation of how good their progress is going on a task based on the bar progression. Now just locking down how users are informed about the hours estimated field, when they first make a new task.